% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brat.R
\docType{package}
\name{brat}
\alias{brat}
\alias{brat-package}
\alias{bratOutput}
\alias{renderBrat}
\title{Create brat htmlwidget}
\usage{
brat(data = list(), width = NULL, height = NULL)

bratOutput(outputId, width = "100\%", height = "400px")

renderBrat(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{data}{Data to be passed into widget.}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{outputId}{The ID of the widget.}

\item{expr}{An expression.}

\item{env}{An environment.}

\item{quoted}{A \code{logical} value.}

\item{x}{Any kind of input.}
}
\description{
This is a lightweight wrapper for the brat annotation tool.
}
\author{
Andreas Blaette
}
