% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brat.R
\docType{data}
\name{example_doc_data}
\alias{example_doc_data}
\alias{example_coll_data}
\title{Brat example data}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 2.
}
\usage{
example_doc_data

example_coll_data
}
\description{
Brat example data

Brat example data
}
\keyword{datasets}
