% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/annfile.R
\name{doc_data}
\alias{doc_data}
\alias{as.doc_data}
\alias{read_doc_data}
\alias{write_ann_file}
\alias{as.Annotation.character}
\title{Get input for brat HTML widget}
\usage{
as.doc_data(x)

read_doc_data(txt_file, ann_file)

write_ann_file(x, ann_file = tempfile(fileext = ".ann"))

as.Annotation.character(ann_file)
}
\arguments{
\item{x}{An object to be transformed.}

\item{txt_file}{Plain text file with fulltext of document.}

\item{ann_file}{A tab-separated file with annotations, brat standoff format.}
}
\description{
Get input for brat HTML widget

Data conversion for brat
}
\examples{
library(NLP)
merkel_min <- merkel

# We do not want to show all annotations
merkel_min$annotation <- merkel$annotation[merkel$annotation$type == "ner"]
d <- as.doc_data(merkel_min)
txt_file <- system.file(package = "brat", "extdata", "sample_data", "edokelley.txt")
ann_file <- system.file(package = "brat", "extdata", "sample_data", "edokelley.ann")
doc_data <- read_doc_data(txt_file, ann_file)
write_ann_file(example_doc_data)
ann_file <- system.file(package = "brat", "extdata", "sample_data", "edokelley.ann")
as.Annotation(ann_file)
}
