% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{entity}
\alias{entity}
\title{Configure entity settings for collection data.}
\usage{
entity(
  type,
  labels = c(type, substr(type, 1, 3)),
  bgColor,
  borderColor = "darken"
)
}
\arguments{
\item{type}{A length-one \code{character} vector defining the type of annotation.}

\item{labels}{A \code{character} vector with two elements: The label to be shown
over an annotation, and an abbreviation. If not stated explicitly, the type
will be used as label and the first three characters of it as an
abbreviation.}

\item{bgColor}{The background color  of the annotation.}

\item{borderColor}{Border color of the annotation. Defaults to "darken".}
}
\value{
A nested \code{list} object.
}
\description{
Collection data require a definition of entities that shall be annotated.
This auxiliary function generates the data format ready to passed to the brat
JavaScript code.
}
\examples{
entity("Person", bgColor = "#7FC97F")
}
