% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brat.R
\docType{data}
\name{merkel}
\alias{merkel}
\title{Annotated Speech of Angela Merkel}
\format{
A \code{AnnotatedPlainTextDocument} object as defined in the NLP package.
}
\usage{
merkel
}
\description{
Annotated Speech of Angela Merkel
}
\keyword{datasets}
