% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brat.R
\docType{data}
\name{millenium_declaration_annotated}
\alias{millenium_declaration_annotated}
\title{Annotated Millenium Declaration}
\format{
A \code{AnnotatedPlainTextDocument} object as defined in the NLP package.
}
\usage{
millenium_declaration_annotated
}
\description{
Annotated Millenium Declaration
}
\keyword{datasets}
