% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relation}
\alias{relation}
\title{Define relation for configuration of collection data.}
\usage{
relation(
  type,
  labels = c(type, substr(type, 1, 3)),
  dashArray = "3,3",
  color,
  roles,
  targets
)
}
\arguments{
\item{type}{The type of relation to be annoated (length-one \code{character} vector).}

\item{labels}{A \code{character} vector with two elements with the label to be shown
for an annotation and an abbreviation of the label that is used when necessary.
Defaults to the \code{type} of the annotation and the first three characters of the
annotation.}

\item{dashArray}{A length-one \code{character} vector defining the visual layout
of arrows.}

\item{color}{The color of the annotation (length-one \code{character} vector).}

\item{roles}{A \code{character} vector with the same length as argument \code{targets}.}

\item{targets}{A \code{character} vector with the same length as argument \code{roles}.}
}
\description{
Annotating relations requires a definition that is part of the collection data.
This auxiliary definition prepares the data structure that can be passed to
brat JavaScript.
}
\examples{
relation("Anaphora", color = "purple", roles = c("Anaphor", "Entity"), targets = c("Person", "Person"))
}
