% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainy.R
\name{brainy}
\alias{brainy}
\title{Brat shiny HTML widget}
\usage{
brainy(doc_dir, coll_data)
}
\arguments{
\item{doc_dir}{Diretory with plaint text documents (\emph{.txt) and annotation
files (}.ann).}

\item{coll_data}{bla bla}
}
\description{
Brat shiny HTML widget
}
\examples{
doc_dir <- system.file(package = "brat", "extdata", "sample_data")

coll_data <- list(
  entity_types = list(
    list(
      type = "Person",
      labels = c("Person", "Per"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[1],
      borderColor = "darken",
      arcs = list(list(targets = list("Person")))
    ),
    list(
      type = "Organization",
      labels = c("Organization", "Org"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[2],
      borderColor = "darken"
    ),
    list(
      type = "Location",
      labels = c("Location", "Loc"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[3],
      borderColor = "darken"
    )
  ),
  relation_types = list(list(
    type = "Anaphora",
    labels = c("Anaphora", "Ana"),
    dashArray = "3,3",
    color = "purple",
    args = list(
      list(role = "Anaphor", targets = "Person"),
      list(role = "Entity", targets = "Person")
    )
   ))
)

txt_src <- system.file(package = "brat", "extdata", "sample_data", "edokelley.txt")
ann_src <- system.file(package = "brat", "extdata", "sample_data", "edokelley.ann")
txt_file <- file.path(tempdir(), "edokelley.txt")
ann_file <- file.path(tempdir(), "edokelley.ann")
file.copy(txt_src, txt_file)
file.copy(ann_src, ann_file)

if (interactive()) brainy(doc_dir = doc_dir, coll_data = coll_data)
if (interactive()){
  brainy(doc_dir = "~/Lab/github/brat/inst/extdata/sample_data", coll_data = coll_data)
}
}
