% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brat.R, R/bradget.R
\docType{package}
\name{brat}
\alias{brat}
\alias{brat-package}
\alias{bratOutput}
\alias{renderBrat}
\alias{bradget}
\title{Create brat htmlwidget}
\usage{
brat(doc_data = list(), coll_data = list(), width = NULL, height = NULL)

bratOutput(outputId, width = "100\%", height = "400px")

renderBrat(expr, env = parent.frame(), quoted = FALSE)

bradget(doc_data, coll_data)
}
\arguments{
\item{doc_data}{bla bla}

\item{coll_data}{bla bla}

\item{width}{The width of the widget.}

\item{height}{The height of the widget.}

\item{outputId}{The ID of the widget.}

\item{expr}{An expression.}

\item{env}{An environment.}

\item{quoted}{A \code{logical} value.}
}
\value{
A \code{AnnotatedPlainTextDocument} object as defined in the NLP package.
}
\description{
This is a lightweight wrapper for the brat annotation tool.
}
\examples{
data <- list(
  text = "Ed O'Kelley was the man who shot the man who shot Jesse James.\n What a guy. He should be hanged.",
  entities = list(
    list('T1', 'Person', list(c(0, 11))),
    list('T2', 'Person', list(c(20, 23))),
    list('T3', 'Person', list(c(37, 40))),
    list('T4', 'Person', list(c(50, 61))),
    list('T5', 'Person', list(c(64, 68)))
  )
)
collData <- list(
  entity_types = list(list(
    type = "Person",
    labels = c("Person", "Per"),
    bgColor = "#7fa2ff",
    borderColor = "darken"
  ))
)
if (interactive()) brat(doc_data = data, coll_data = collData)

# A second example

library(NLP)
merkel_min <- merkel
merkel_min$annotation <- merkel$annotation[merkel$annotation$type == "ner"]
d <- as.BratDocData(merkel_min)
collData <- list(
  entity_types = list(list(
    type = "ner",
    labels = c("Named Entity", "NE"),
    bgColor = "#7fa2ff",
    borderColor = "darken"
  ))
)
if (interactive()) brat(doc_data = d, coll_data = collData)
doc_data <- list(
  text = "Ed O'Kelley was the man who shot the man who shot Jesse James.\n What a guy. He should be hanged.",
  entities = list(
    list('T1', 'Person', list(c(0, 11))),
    list('T2', 'Person', list(c(20, 23))),
    list('T3', 'Person', list(c(37, 40))),
    list('T4', 'Person', list(c(50, 61))),
    list('T5', 'Person', list(c(64, 68)))
  )
)
coll_data <- list(
  entity_types = list(
    list(
      type = "Person",
      labels = c("Person", "Per"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[1],
      borderColor = "darken"
    ),
    list(
      type = "Organization",
      labels = c("Organization", "Org"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[2],
      borderColor = "darken"
    ),
    list(
      type = "Location",
      labels = c("Location", "Loc"),
      bgColor = RColorBrewer::brewer.pal(8, "Accent")[3],
      borderColor = "darken"
    )
  )
)
if (interactive()) a <- bradget(doc_data = doc_data, coll_data = coll_data)

# A second example 

library(NLP)
merkel_min <- merkel
merkel_min$annotation <- merkel$annotation[merkel$annotation$type == "ner"]
d <- as.BratDocData(merkel_min)
collData <- list(
  entity_types = list(list(
    type = "ner",
    labels = c("Named Entity", "NE"),
    bgColor = "#7fa2ff",
    borderColor = "darken"
  ))
)
if (interactive()) bradget(doc_data = d, coll_data = collData)
}
\author{
Andreas Blaette
}
