% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{doc_data}
\alias{doc_data}
\alias{as.doc_data}
\title{Get input for brat HTML widget}
\usage{
as.doc_data(x)
}
\arguments{
\item{x}{An object to be transformed.}
}
\description{
Get input for brat HTML widget
}
\examples{
library(NLP)
merkel_min <- merkel

# We do not want to show all annotations
merkel_min$annotation <- merkel$annotation[merkel$annotation$type == "ner"]
d <- as.doc_data(merkel_min)
}
